reloadResources() once, for packs, on update.
Found while using the mod that it triggered abnormally when used in
combination with respackopts https://git.jfronny.dev/JfMods/Respackopts,
so added a file path matcher to reject non-.zip files in the resource
location. I don’t know if that is 100% fool-proof, it might require
other extensions to be included? No idea if there is a popular common
name.

HotReloader also fires reloadResources() twice on this Linux system for
any pack modification, so added a short (1.5s) cooldown to prevent
chained events for triggering again too soon.